+++ src/main/java/org/javlo/helper/URLHelper.java
@@ -520,6 +520,10 @@ public class URLHelper extends ElementaryURLHelper {
 
 		return createStaticURL(ctx, null, URLHelper.mergePath(templateFullPath, url), true);
 	}
+	
+	public static String createActionURL(ContentContext ctx, String action, String inUrl) { 
+		return createStaticURL(ctx, mergePath("webaction",action,inUrl));
+	}
 
 	public static String createStaticURL(ContentContext ctx, MenuElement referencePage, String inUrl) {
 		return createStaticURL(ctx, referencePage, inUrl, true);